Atualizando 12/08

+++ addfotos.php
@@ -9,35 +9,46 @@
 	</header>
 	<body>
 		<?php include('conn.php'); date_default_timezone_set('UTC');?>
-		
-		<table border="2">
-			<tr>
-				<th>#</th>
-				<th>Foto</th>
-				<th>Nome</th>
-				<th>Ação</th>
-			</tr>
-			<?php
-				$sql = "SELECT * from tb_galeria gal";
-				$result = $mysqli->query($sql);
-				if($result->num_rows > 0){
-					while($row = $result->fetch_object()){
-						echo
-							"<tr>
-								<td>".$row->cd_foto."</td>
-								<td><img class='mini' src='".$row->ds_endereco."' title='".substr($row->ds_endereco, 8)."'></td>
-								<td>".substr($row->ds_endereco, 8)."</td>
-								<td><a href='addfotos.php?codfoto=".$row->cd_foto."'><button id='apagar'>Apagar</button></a></td>
-							</tr>";
+		<form method="POST">
+			<table border="2">
+				<tr>
+					<th>#</th>
+					<th>Selecionar</th>
+					<th>Foto</th>
+					<th>Nome</th>
+					<th>Ação</th>
+					
+				</tr>
+				<?php
+					$sql = "SELECT * from tb_galeria gal";
+					$result = $mysqli->query($sql);
+					if($result->num_rows > 0){
+						$c = 0;
+						while($row = $result->fetch_object()){
+							$c++;
+							echo
+								"<tr>
+									<td>".$row->cd_foto."</td>
+									<td align='center'><label for='fotos".$c."'><input type='checkbox' name='fotos[]' id='fotos".$c."' class='fotos filled-in' value='".$row->cd_foto."'><span></span></label></td>
+									<td><label for='fotos".$c."'><img class='mini' src='".$row->ds_endereco."' title='".substr($row->ds_endereco, 8)."'></label></td>
+									<td>".substr($row->ds_endereco, 8)."</td>
+									<td><a href='addfotos.php?codfoto=".$row->cd_foto."'><button id='apagar'>Apagar</button></a></td>
+									
+								</tr>";
+						}
 					}
-				}
-				else{
-					echo "<tr><td></td><td></td><td>Nenhuma foto adicionada.</td><td></tr>";
-				}
-			?>
-		</table>
-		<br><center><a href="addfotos.php"><button>Atualizar</button></a></center>	
-		<form  action="addfotos.php" method="post" multipart="" enctype="multipart/form-data">
+					else{
+						echo "<tr><td></td><td></td><td>Nenhuma foto adicionada.</td><td></tr>";
+					}
+				?>
+			</table>
+		
+		<br>
+		<center>
+			<a href="addfotos.php"><button>Atualizar</button></a>
+			<button id="excluir" type="submit">Excluir selecionados</button></form>
+		</center>	
+		<form action="addfotos.php" method="post" multipart="" enctype="multipart/form-data">
 			<label for="img">
 				<b>Adicionar fotos: </b><input type="file" name="img[]" id="img" multiple>
 			</label>
@@ -67,14 +78,48 @@
 			}
 			if(isset($_GET['codfoto'])){
 				$foto = $_GET['codfoto'];
-				$sql = "DELETE from tb_galeria where cd_foto = '$foto'";
-				if(!$mysqli->query($sql)){
-					echo "Error: " . $sql . "<br>" . mysqli_error($mysqli);
+				$sql = "SELECT * from tb_galeria where cd_foto = '$foto'";
+				$result = $mysqli->query($sql);
+				if($result->num_rows > 0){
+					$row = $result->fetch_object();
+					$arquivo = $row->ds_endereco;
+					$sql = "DELETE from tb_galeria where cd_foto = '$foto'";
+					if(!$mysqli->query($sql)){
+						echo "Error: " . $sql . "<br>" . mysqli_error($mysqli);
+					}
+					else{
+						if(unlink($arquivo)){
+							echo "<script type='text/javascript'>window.location.href='addfotos.php';</script>";
+						}
+					}
 				}
-				else{
-					echo "<script type='text/javascript'>window.location.href='addfotos.php';</script>";
+			}
+			if(isset($_POST['fotos'])) {
+				foreach ($_POST['fotos'] as $foto){
+					$sql = "SELECT * from tb_galeria where cd_foto = '$foto'";
+					$result = $mysqli->query($sql);
+					if($result->num_rows > 0){
+						$row = $result->fetch_object();
+						$arquivo = $row->ds_endereco;
+						$sql = "DELETE from tb_galeria where cd_foto = '$foto'";
+						if(!$mysqli->query($sql)){
+							echo "Error: " . $sql . "<br>" . mysqli_error($mysqli);
+						}
+						else{
+							unlink($arquivo);
+						}
+					}
 				}
+				echo "<script type='text/javascript'>window.location.href='addfotos.php';</script>";
 			}
 		?>
 	</body>
+	<script>
+		$(document).ready(function(){
+			$("#excluir").hide();
+			$(".fotos").click(function(){
+				$("#excluir").show();
+			});
+		});
+	</script>
 </html>
